create libft dir, add some more functions, delete wrong Makefile which somehow appear in src
