+++ src/libft/ft_strchr.c
@@ -1,0 +1,17 @@
+#include "minishell.h"
+
+char	*ft_strchr(const char *s, int c)
+{
+	int	i;
+
+	i = 0;
+	while (s[i])
+	{
+		if (s[i] == (char)c)
+			return ((char *)(s + i));
+		i++;
+	}
+	if (s[i] == (const char)c)
+		return ((char *)(s + i));
+	return (NULL);
+}
