+++ src/libft/ft_strtrim.c
